Created new Features: Seeders, Factories and change the models

+++ backend/app/Http/Controllers/Api/ProjectController.php
@@ -17,6 +17,7 @@ class ProjectController extends Controller
 
     public function store(ProjectRequest $request)
     {
+        $this->authorize('create', Project::class);
         $project = Project::create($request->validated());
         return response()->json([
             'success' => true,
@@ -24,14 +25,15 @@ class ProjectController extends Controller
             'data' => $project
         ], 201);
     }
-
+    
     public function show(Project $project)
     {
         return new ProjectResource($project);
     }
-
+    
     public function update(ProjectRequest $request, Project $project)
     {
+        $this->authorize('update', $project);
         $project->update($request->validated());
         return response()->json([
             'success' => true,
@@ -42,6 +44,7 @@ class ProjectController extends Controller
 
     public function destroy(Project $project)
     {
+        $this->authorize('delete', $project);
         $project->delete();
         return response()->json(null, 204);
     }
